+++ libraries/shared/src/PerfStat.cpp
@@ -105,7 +105,7 @@ int PerfStat::DumpStats(char** array) {
 
 // Destructor handles recording all of our stats
 PerformanceWarning::~PerformanceWarning() {
-    long long end = usecTimestampNow();
+    uint64_t end = usecTimestampNow();
     double elapsedmsec = (end - _start) / 1000.0;
     if ((_alwaysDisplay || _renderWarningsOn) && elapsedmsec > 1) {
         if (elapsedmsec > 1000) {
