+++ voxel-server/src/main.cpp
@@ -32,7 +32,7 @@
 
 const char* LOCAL_VOXELS_PERSIST_FILE = "resources/voxels.svo";
 const char* VOXELS_PERSIST_FILE = "/etc/highfidelity/voxel-server/resources/voxels.svo";
-const long long VOXEL_PERSIST_INTERVAL = 1000 * 30; // every 30 seconds
+const int VOXEL_PERSIST_INTERVAL = 1000 * 30; // every 30 seconds
 
 const int VOXEL_LISTEN_PORT = 40106;
 
@@ -119,7 +119,7 @@ void resInVoxelDistributor(NodeList* nodeList,
     bool searchReset = false;
     int  searchLoops = 0;
     int  searchLevelWas = nodeData->getMaxSearchLevel();
-    long long start = usecTimestampNow();
+    uint64_t start = usecTimestampNow();
     while (!searchReset && nodeData->nodeBag.isEmpty()) {
         searchLoops++;
 
@@ -138,7 +138,7 @@ void resInVoxelDistributor(NodeList* nodeList,
             }
         }
     }
-    long long end = usecTimestampNow();
+    uint64_t end = usecTimestampNow();
     int elapsedmsec = (end - start)/1000;
     if (elapsedmsec > 100) {
         if (elapsedmsec > 1000) {
@@ -162,7 +162,7 @@ void resInVoxelDistributor(NodeList* nodeList,
         int packetsSentThisInterval = 0;
         int truePacketsSent = 0;
         int trueBytesSent = 0;
-        long long start = usecTimestampNow();
+        uint64_t start = usecTimestampNow();
 
         bool shouldSendEnvironments = shouldDo(ENVIRONMENT_SEND_INTERVAL_USECS, VOXEL_SEND_INTERVAL_USECS);
         while (packetsSentThisInterval < PACKETS_PER_CLIENT_PER_INTERVAL - (shouldSendEnvironments ? 1 : 0)) {
@@ -209,7 +209,7 @@ void resInVoxelDistributor(NodeList* nodeList,
             trueBytesSent += envPacketLength;
             truePacketsSent++;
         }
-        long long end = usecTimestampNow();
+        uint64_t end = usecTimestampNow();
         int elapsedmsec = (end - start)/1000;
         if (elapsedmsec > 100) {
             if (elapsedmsec > 1000) {
@@ -248,7 +248,7 @@ void deepestLevelVoxelDistributor(NodeList* nodeList,
     pthread_mutex_lock(&::treeLock);
 
     int maxLevelReached = 0;
-    long long start = usecTimestampNow();
+    uint64_t start = usecTimestampNow();
 
     // FOR NOW... node tells us if it wants to receive only view frustum deltas
     bool wantDelta = nodeData->getWantDelta();
@@ -267,7 +267,7 @@ void deepestLevelVoxelDistributor(NodeList* nodeList,
         if (::debugVoxelSending) {
             printf("(viewFrustumChanged=%s || nodeData->nodeBag.isEmpty() =%s)...\n",
                    debug::valueOf(viewFrustumChanged), debug::valueOf(nodeData->nodeBag.isEmpty()));
-            long long now = usecTimestampNow();
+            uint64_t now = usecTimestampNow();
             if (nodeData->getLastTimeBagEmpty() > 0) {
                 float elapsedSceneSend = (now - nodeData->getLastTimeBagEmpty()) / 1000000.0f;
                 
@@ -306,7 +306,7 @@ void deepestLevelVoxelDistributor(NodeList* nodeList,
         }
 
     }
-    long long end = usecTimestampNow();
+    uint64_t end = usecTimestampNow();
     int elapsedmsec = (end - start)/1000;
     if (elapsedmsec > 100) {
         if (elapsedmsec > 1000) {
@@ -329,12 +329,12 @@ void deepestLevelVoxelDistributor(NodeList* nodeList,
         int packetsSentThisInterval = 0;
         int truePacketsSent = 0;
         int trueBytesSent = 0;
-        long long start = usecTimestampNow();
+        uint64_t start = usecTimestampNow();
 
         bool shouldSendEnvironments = shouldDo(ENVIRONMENT_SEND_INTERVAL_USECS, VOXEL_SEND_INTERVAL_USECS);
         while (packetsSentThisInterval < PACKETS_PER_CLIENT_PER_INTERVAL - (shouldSendEnvironments ? 1 : 0)) {        
             // Check to see if we're taking too long, and if so bail early...
-            long long now = usecTimestampNow();
+            uint64_t now = usecTimestampNow();
             long elapsedUsec = (now - start);
             long elapsedUsecPerPacket = (truePacketsSent == 0) ? 0 : (elapsedUsec / truePacketsSent);
             long usecRemaining = (VOXEL_SEND_INTERVAL_USECS - elapsedUsec);
@@ -396,7 +396,7 @@ void deepestLevelVoxelDistributor(NodeList* nodeList,
             truePacketsSent++;
         }
         
-        long long end = usecTimestampNow();
+        uint64_t end = usecTimestampNow();
         int elapsedmsec = (end - start)/1000;
         if (elapsedmsec > 100) {
             if (elapsedmsec > 1000) {
@@ -425,10 +425,10 @@ void deepestLevelVoxelDistributor(NodeList* nodeList,
     pthread_mutex_unlock(&::treeLock);
 }
 
-long long lastPersistVoxels = 0;
+uint64_t lastPersistVoxels = 0;
 void persistVoxelsWhenDirty() {
-    long long now = usecTimestampNow();
-    long long sinceLastTime = (now - ::lastPersistVoxels) / 1000;
+    uint64_t now = usecTimestampNow();
+    int sinceLastTime = (now - ::lastPersistVoxels) / 1000;
 
     // check the dirty bit and persist here...
     if (::wantVoxelPersist && ::serverTree.isDirty() && sinceLastTime > VOXEL_PERSIST_INTERVAL) {
@@ -473,7 +473,7 @@ void *distributeVoxelsToListeners(void *args) {
         }
         
         // dynamically sleep until we need to fire off the next set of voxels
-        long long usecToSleep =  VOXEL_SEND_INTERVAL_USECS - (usecTimestampNow() - usecTimestamp(&lastSendTime));
+        int usecToSleep =  VOXEL_SEND_INTERVAL_USECS - (usecTimestampNow() - usecTimestamp(&lastSendTime));
         
         if (usecToSleep > 0) {
             usleep(usecToSleep);
